+++ components/renderers/cpp/src/splatter_cuda.hpp
@@ -141,7 +141,7 @@ namespace cuda {
 
 	void show_mask(
         ftl::cuda::TextureObject<uchar4> &colour,
-		ftl::cuda::TextureObject<int> &mask,
+		ftl::cuda::TextureObject<uint8_t> &mask,
         int id, uchar4 style, cudaStream_t stream);
 
 	void merge_convert_depth(
